+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Add `sdwan_multicast_feature_template` resource and data source
 - Add `sdwan_ntp_feature_template` resource and data source
 - Add `sdwan_archive_feature_template` resource and data source
+- Add `sdwan_banner_feature_template` resource and data source
 
 ## 0.3.6